feat: add support for annotations
These annotations are magic comments that can be used by the analysis
passes to provide additional information. This commit adds the general
framework for annotations, and adds support for:
```clarity
  ;; #[allow(unchecked_data)]
```
which can be used to tell the check-checker to ignore an expression that
would otherwise trigger an unchecked data warning. The syntax is similar
to Rust attributes, as is the functionality. This annotation applies to
the expression that follows the annotation. See unit tests for examples.