fix: handle private functions in check-checker
All arguments to private functions are required to be checked (not
tainted). This keeps the checking of the untrusted data close to its
source, in the public function.

A new annotation, `allow(unchecked_params)` is added to allow a
developer to specify that a private function is allowed to receive
unchecked arguments. This currently applies to all parameters, but
should be enhanced to specify which parameters.