feat: add 'filter' annotation
The filter annotation allows developers to tell the check-checker to
filter one or more inputs on an expression that cannot be detected as a
filter. This is useful when a check on some unrelated inputs indirectly
validates another input. For example:

```clarity
(define-public (withdraw (amount uint))
    (begin
        ;; #[filter(amount)]
        (asserts! (is-eq tx-sender (var-get admin)) (err u400))
        (stx-transfer? amount (as-contract tx-sender) tx-sender)
    )
)
```

The filter annotation takes either a comma-separated list of parameter
names to filter, or a '*', which filters everything.

Resolves #72