fix: crash when an error is reported at EOF
When there is a newline at the end of a file, and an error is reported
at the very end (for example, for a missing ')'), this would crash when
attempting to print out the line. This is fixed by only advancing the
line number when the next token is not the end of the file.