feat: improve check-checker handling of rollbacks
When a function returns an `err` type, any mutations on the blockchain
state from that function, or its callees, are aborted. This means that
if a value is checked after it is used, and the check triggers an error
to be returned from the function (such as with an `asserts!`), then the
original use should no longer be considered unchecked.

Resolves: #81