fix: allow symbols in identifiers
To match the current implementation in the blockchain, identifiers
should be allowed to have special characters like `*`, `/`, `<`, etc. in
them, although, a future specification of the Clarity language may make
those illegal. Now, the parser will report a warning for those cases.

This change required some modified handling of the trait identifiers
(ex. `<my-trait>`), so the changes are more significant than might be
expected.