fix(debugger): handle breaking in debug command
If the expression included in the `::debug` command includes
sub-expressions, then the debugger will begin by breaking at the first
sub-expression. Previously, the debugger was unable to retrieve the
source code for this expression to print it in the console. This fixes
that by saving it in the `DebugState` object and checking for this case.