fix: harden `ASTVisitor` to handle invalid ASTs
Originally, `ASTVisitor` was only used after the type checker, so it could
be assumed that the AST was well-formed. Now, we are using it very
early, to detect dependencies in the AST. This changeset avoids crashing
by replacing all prior uses of `unwrap()` with `unwrap_or` or
`unwrap_default`, so that a placeholder is used in the case where the
expected type is not present.

Fixes: hirosystems/clarinet#334