fix: check the maximum length of a contract name
The parser used `ClarityName::try_from` to check for a valid name, but
this check did not check the length of the name (must be 1-40
characters). Also perform a check in the parser for other names that are
too long, rather than relying on the `try_from` to catch it.

Fixes: #145