feat: allow pasting of multiple lines
This fixes an interactive clarinet bug, and allows for pasting multiple
lines into the console.  I've tested it lightly by copying/pasting
several lines from another window into clarinet console.

There are a couple of changes in behavior.

* The console echos back the form before its result:

    >> (concat "a " " b") (concat "r" "d")
    ;; (concat "a " " b")
    "a  b"
    ;; (concat "r" "d")
    "rd"
    >>

* Multiple expressions on the same line end up as separate entries
  in the history.  Pressing the up arrow after the above input
  brings up only the second form, not both of them.  This might not
  be desirable.

* This also works, though i'm not sure if it should:

    (concat "a " " b")(concat "r" "d")

Without my change, this results in

    >> (concat "a " " b")(concat "r" "d")
    error: Expected whitespace or a close parens. Found: '('
    warning: conflict between parser versions, reverting to parser v1
    Help improve clarinet by sharing the code that triggered this warning:
    https://github.com/hirosystems/clarinet/issues/new/choose

The bug: the complete_input function did not check if the parens
were inside strings before trying to match them.  This is valid
code that it rejected:

    (concat "abc" "de}f")

Fixes: #115