fix: limit scope of stacks_rw_db usage to free lock (#543)
Every 32 stacks blocks, Chainhook attempts to check for a new archive
file to fix any potential holes in the database. This process requires
readwrite access to the stacks.rocksdb. However, the process that called
this subprocess already had a readwrite instance of the database open,
which was preventing the lock from being cleared.

This PR sets the scope of the initial usage of the db in order to free
up the lock.