fix: make bitcoin payload serialization deserializable (#569)
When a Chainhook is triggered and we send a payload to a user, we filter
some values based on user settings (`include_proof`, `include_inputs`,
etc). This helps to reduce the payload size that is sent over the wire.

However, when we created this payload, we were omitting these fields
completely even though the corresponding type does not have those fields
as optional. This makes using our SDK types for deserialization
impossible.

This PR adds in those fields in all cases and sets them to empty values
if filtered out.