fix: allow aborting a predicate scan (#601)
Previously, when Chainhook was run as a service and the runloop to scan
stacks/bitcoin predicates was set, we had no way to abort that scan. If
a predicate was set to scan 1m blocks, but the user discovered the
predicate was wrong and needed to delete, the user could delete the
predicate from the store, but the scan thread had already started and
would run until completion.

This PR adds an abort signal to an ongoing scan so that when a predicate
is deregistered, the scan is canceled.